% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/head_df.R
\name{head_df}
\alias{head_df}
\alias{tail_df}
\title{Return the first or the last part of a data.frame}
\usage{
head_df(x, pos = "left", nrow = 3, ncol = 4)

tail_df(x, pos = "left", nrow = 3, ncol = 4)
}
\arguments{
\item{x}{data frame or matrix-like object.}

\item{pos}{"left" or "right"}

\item{nrow}{numeric. number of rows.}

\item{ncol}{numeric. number of columns.}
}
\value{
returns a data-frame or a matrix-like object with head or tail of the data-frame.
}
\description{
\code{head_df} returns top-left (when \code{pos = 'left'}) or top-right (when \code{pos = 'right'}) part of a data.frame or a matrix-like object.
Similarly, \code{tail_df} returns bottom-left or bottom-right part of a data frame.
}
\examples{
x = matrix(1:100, nrow = 10, byrow = F, dimnames = list(paste0('r',1:10), paste0('c',1:10)))
head_df(x)
head_df(x, 'right')
tail_df(x, nrow = 2)
tail_df(x, 'right', nrow = 5, ncol = 5)
}
