% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_df.R
\name{split_df}
\alias{split_df}
\title{Function to split data.}
\usage{
split_df(x, frac, by.col = T)
}
\arguments{
\item{x}{data.frame or similar object.}

\item{frac}{numeric vector. fraction of data in different parts. if \code{sum(frac)!=1}, thne \code{frac} will be normalized.}

\item{by.col}{logical. If True, split columns, otherwise split rows.}
}
\value{
list of data frames with splitted data. 
if \code{frac} vector has names, so does the returned list.
}
\description{
\code{split_df} splits a data frame or similar object randomly with specified fractions of columns (or rows) in each split.
}
\examples{
x = matrix(1:50, nrow = 5, dimnames = list(paste0('r',1:5),paste0('c',1:10)))
splitted_x = split_df(x, frac = c(train=0.6, validation=0.2, test=0.2))
}
